requirements.txt
bert-for-tf2
tensorflow>=2.0.0
